Reduce module load attempts
When using the Rollbar logger, we've noticed a lot of noise in strace
attempting to the load the request objects from various web frameworks.

Since the request object is only used in the final report sent to
Rollbar, only attempt to get it when the code is sure it wants to send a
report.

Use the import checks already done for request objects to skip import
attempts for frameworks that can't possibly be available, as their base
request objects are not available.