Fix memory leak in in history collection
History is stored as a list of logger records. After determining the
history for a given record, that history was then added to the record.
This led to a tree of records. Even though the history in the logger was
truncated, active records in the history would continue to point to
inactive ones, and so on, causing a memory leak.

Since the history field on record objects was only used within the
_add_history method, convert it to a local variable, and kill the leak.