Fix JSON encoding and update to better support Python 2/3
JSON encoding fails sometimes if there is a local variable or custom data
passed in that cannot be converted to unicode and then to utf8 encoded
bytes. This change implements a custom iterencode() method for the JSON
encoder. It takes each chunk of JSON that is produced during JSON serialization
and attempts to decode it using UTF-8 encoding. If a decode error occurs,
a custom unicode string will be added to the JSON output that gives
information about the decode error and the data that produced it.

Needs more testing.

@brianr