More fixes for python version differences
Python3 bytes need to be handled with the transform_string() method.
Added force_lower() which will attempt to turn whatever it's passed
into a lowercased value even if it needs to typecast it.
Added a map() helper to fix inconcistent implementations for Python 2/3.
Fixed a few tests that treated Python 2/3 results differently but
checked the Python version incorrectly.