Simplified circular reference handling
Moved the logic for types that are allowed to be circularly
referenced into the traverse() algorithm. This makes things
much simpler since we no longer have to worry about the
correct type handling in the transform.transform_circular_reference()
method.

Fixed a bug in the URL scrubbing logic which was scrubbing
non-strings if the suffix matched. Added a test for this.

Finally, based on Brian's feedback, changed the format of
the circular reference label.

@brianr