Fixes bug that caused some payload fields to become invalid
The problem was that the serializer was being run on the final payload
object which had references in the request dict to local variables in
the trace frames. This caused request.params = <CircularReference ...>
which is invalid according to the Rollbar API schema.

This change serializes the payload in sections so that each section
cannot reference the other while checking for circular references.

@brianr