revert exception safe repr
This reverts a pull request we accepted and merged recently
(https://github.com/rollbar/pyrollbar/pull/91).  It turns out that that
request causes excpetion messages to not be passed along as expected,
and that isn't desired behaviour.  We're going to attack this problem
again soon and come up with a solution that meets the needs of the
reverted pull request without causing the undesired side effects.