In Pyramid module, raise tween above EXCVIEW.
Currently the Rollbar tween is put below the EXCVIEW.  It's desirable to
put the Rollbar tween above EXCVIEW to catch errors that might occur
within EXCVIEW.  Also, in order to catch errors in tweens that are
placed above EXCVIEW (such as `pyramid_tm`), Rollbar must support being
placed above EXCVIEW as well.

This commit modifies the Rollbar tween to extract an exception from
`request.exception` and `request.exc_info` if available.  It also moves
the tween to above EXCVIEW.