fix deque test's expected repr()
In Python > 3.5 the deque class is registered as a MutableSequence
https://github.com/python/cpython/blob/3.5/Lib/collections/__init__.py#L39,
and this is already handled as a special case. This also affects Python
2.7 if the backported "typing" module is imported. As it becomes more
and more likely a Python library on PyPI is typed, this is likely to
continue to be a moving target.

Currently, it looks like the twisted test case relies on treq, which
relies on hyperlink, which will import the typing module.