Add consuming request body in Starlette middleware
Due to bug in Starlette, consuming request body in middlewares can
freeze application. This usually happen for FastAPI apps as the
framework tries to consume the data itself.

This commit adds consuming request body in Starlette middleware,
however it's disabled by default (added as a comment).
In case of uncommenting the body consumption, the payload builder
verifies if body is available and append it to the payload.

See: https://github.com/encode/starlette/issues/495#issuecomment-494008175