Use local context when using async_handler ctx manager
When reporting via concurrent coroutines, the async_handler ctx manager
stores a locally substituted async handler. This allows each coroutine
to use different handler.

This solution is based on `contextvars` module that was added in
Python 3.7+. In case of Python 3.6, it tries to backport the module
using `aiocontextvars` 3rd party package.